[TASK] Eliminate "generated" headlines
In the OpenOffice document we sometimes left out one level of headlines.
During the migration to ReST, however every level of headlines, if it
existed or not, was written in the .rst files; also the missing ones.
These headlines now show up as "((generated))" in the files and in the
rendered output.

Replace each "generated" headline with the one which should actually be
there.

Change-Id: Ia53ec07e5c2c72df6d7552206874b4b3ced508a2
Resolves: #47952
Reviewed-on: https://review.typo3.org/20540
Reviewed-by: Christopher
Tested-by: Christopher
Reviewed-on: https://review.typo3.org/20624